added concept checks to some xexpression functions
